RAPLDrv: use METHOD_BUFFERED to transfer data from userland to kernelland and vice-versa
This patch disables old used method (where data is only transfered from driver to user).

This patch attempt to read a MSR register index value passed by the user (as a 4 bytes in length input), converted into an UINT32 value and then is passed to the '__readmsr' intrinsic function.
For testing purposes, MSR register indexes of value 1, 2 and 3 returns the following values:
 - 1: 0x1111111111111111
 - 2: 0x2222222222222222
 - 3: part of Kernel system time (8 bytes length)

TODO: test by passing a random value which is not a value MSR register index and see what happens (BSOD?)