Agent: Add manufacturer code for both x86_64 and x86 targets. Also includes RAPL defines from Linux headers.
NOTES: as x86_64 assembly code cannot be compiled with MSVC tools, the user has to link the Agent against the generated NASM object.
For x86 target, the assembly code is directly inlined and the user does not have to link the Agent against the generated NASM object (of course he can do it it would work).

This commit also demonstrates the usage of Intel or AMD RAPL values depending on the CPU manufacturer.