Driver: secure MSRs access with try/except combination
Some MSRs that Intel's documentation says it's available, aren't.
An example of that is the MSR register 0x641 (MSR_PP1_ENERGY_STATUS) which
is available for the Haswell CPU series.
Accessing this register on a Intel Xeon E5-2630v3 CPU results in a unhandled kernel
exception which crashes the whole system because this particular CPU does not
have an integrated graphics chipset (as well as all Xeon CPU variant).

In order to avoid this behavior, secure MSRs access (via rdmsr instruction)
with a try/except combination.
If the exception occurs, the userland has to check the driver reply using
the 'GetLastError' API and check if the error code is STATUS_IO_DEVICE_ERROR.