Driver: fix CPU model detection in MSR validation
Previously, the functions defined in msr.c used a different copy of the machine_type
variable than those in driver.c, resulting in always incorrectly assuming an Intel CPU
even when an AMD CPU was successfully detected. By switching to an extern reference to
the static variable used by the driver, the CPU mode is again detected successfully.

Co-authored-by: Victorien Molle <biche@biche.re>
Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>